+++ axum/src/routing/tests/fallback.rs
@@ -1,4 +1,5 @@
 use super::*;
+use crate::middleware::{map_request, map_response};
 
 #[tokio::test]
 async fn basic() {
@@ -58,4 +59,103 @@ async fn fallback_accessing_state() {
     let res = client.get("/does-not-exist").send().await;
     assert_eq!(res.status(), StatusCode::OK);
     assert_eq!(res.text().await, "state");
+}
+
+async fn inner_fallback() -> impl IntoResponse {
+    (StatusCode::NOT_FOUND, "inner")
+}
+
+async fn outer_fallback() -> impl IntoResponse {
+    (StatusCode::NOT_FOUND, "outer")
+}
+
+#[tokio::test]
+async fn nested_router_inherits_fallback() {
+    let inner = Router::new();
+    let app = Router::new().nest("/foo", inner).fallback(outer_fallback);
+
+    let client = TestClient::new(app);
+
+    let res = client.get("/foo/bar").send().await;
+    assert_eq!(res.status(), StatusCode::NOT_FOUND);
+    assert_eq!(res.text().await, "outer");
+}
+
+#[tokio::test]
+async fn doesnt_inherit_fallback_if_overriden() {
+    let inner = Router::new().fallback(inner_fallback);
+    let app = Router::new().nest("/foo", inner).fallback(outer_fallback);
+
+    let client = TestClient::new(app);
+
+    let res = client.get("/foo/bar").send().await;
+    assert_eq!(res.status(), StatusCode::NOT_FOUND);
+    assert_eq!(res.text().await, "inner");
+}
+
+#[tokio::test]
+async fn deeply_nested_inherit_from_top() {
+    let app = Router::new()
+        .nest("/foo", Router::new().nest("/bar", Router::new()))
+        .fallback(outer_fallback);
+
+    let client = TestClient::new(app);
+
+    let res = client.get("/foo/bar/baz").send().await;
+    assert_eq!(res.status(), StatusCode::NOT_FOUND);
+    assert_eq!(res.text().await, "outer");
+}
+
+#[tokio::test]
+async fn deeply_nested_inherit_from_middle() {
+    let app = Router::new().nest(
+        "/foo",
+        Router::new()
+            .nest("/bar", Router::new())
+            .fallback(outer_fallback),
+    );
+
+    let client = TestClient::new(app);
+
+    let res = client.get("/foo/bar/baz").send().await;
+    assert_eq!(res.status(), StatusCode::NOT_FOUND);
+    assert_eq!(res.text().await, "outer");
+}
+
+#[tokio::test]
+async fn with_middleware_on_inner_fallback() {
+    async fn never_called<B>(_: Request<B>) -> Request<B> {
+        panic!("should never be called")
+    }
+
+    let inner = Router::new().layer(map_request(never_called));
+    let app = Router::new().nest("/foo", inner).fallback(outer_fallback);
+
+    let client = TestClient::new(app);
+
+    let res = client.get("/foo/bar").send().await;
+    assert_eq!(res.status(), StatusCode::NOT_FOUND);
+    assert_eq!(res.text().await, "outer");
+}
+
+#[tokio::test]
+async fn also_inherits_default_layered_fallback() {
+    async fn set_header<B>(mut res: Response<B>) -> Response<B> {
+        res.headers_mut()
+            .insert("x-from-fallback", "1".parse().unwrap());
+        res
+    }
+
+    let inner = Router::new();
+    let app = Router::new()
+        .nest("/foo", inner)
+        .fallback(outer_fallback)
+        .layer(map_response(set_header));
+
+    let client = TestClient::new(app);
+
+    let res = client.get("/foo/bar").send().await;
+    assert_eq!(res.status(), StatusCode::NOT_FOUND);
+    assert_eq!(res.headers()["x-from-fallback"], "1");
+    assert_eq!(res.text().await, "outer");
 }