+++ axum/CHANGELOG.md
@@ -9,7 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - **added:** Add `accept_unmasked_frames` setting in WebSocketUpgrade ([#1529])
 - **added:** Add `WebSocketUpgrade::on_failed_upgrade` to customize what to do
-  when upgrading a connection fails
+  when upgrading a connection fails ([#1539])
+
+[#1539]: https://github.com/tokio-rs/axum/pull/1539
 
 # 0.6.0-rc.4 (9. November, 2022)