+++ axum/CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - **fixed:** Nested routers will now inherit fallbacks from outer routers ([#1521])
 - **added:** Add `accept_unmasked_frames` setting in WebSocketUpgrade ([#1529])
+- **added:** Add `WebSocketUpgrade::on_failed_upgrade` to customize what to do
+  when upgrading a connection fails ([#1539])
+
+[#1539]: https://github.com/tokio-rs/axum/pull/1539
 
 [#1521]: https://github.com/tokio-rs/axum/pull/1521